backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
supabase
google-generativeai
pymupdf
pydantic-settings